portfolio/requirements.txt
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1
click==8.1.7
watchdog==3.0.0

# Development dependencies
ruff==0.6.7
pre-commit==3.8.0
